Refactor BookItem
