+++ frontend/src/pages/index.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head'
 import { Inter } from 'next/font/google'
 import styles from '@/styles/Home.module.css'
 import Navbar from '@/components/navbar'
-import BookGrid from '@/components/book'
+import BookGrid from '@/components/BookGrid'
 import Book from '@/entities/book'
 
 const inter = Inter({ subsets: ['latin'] })
